pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic Retrieval for in RAG pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chonkie[hub]>=1.0.7",
    "fastembed>=0.7.0",
    "fastparquet>=2024.11.0",
    "joblib>=1.5.0",
    "langchain-openai>=0.3.17",
    "langchain>=0.3.25",
    "langchain-core>=0.3.60",
    "langgraph>=0.4.5",
    "loguru>=0.7.3",
    "markitdown[all]>=0.1.1",
    "openai>=1.79.0",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "turbopuffer[fast]>=0.4.1",
    "cohere>=5.15.0",
    "judges>=0.0.10",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.3.1",
    "faiss-cpu>=1.9.0",
    "rank-bm25>=0.2.2",
    "transformers>=4.48.0",
    "pdbpp>=0.10.3",
    "groq>=0.13.0",
    "langchain-groq>=0.2.0",
    "streamlit>=1.38.0",
]

[tool.ruff]
line-length = 120
